# api/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Developer Portal (ADP) API
  version: 1.0.0-alpha
  description: |
    REST API for AI agent governance, audit trails, and documentation.
    Supports both PostgreSQL and SQLite backends.

    ## Overview

    ADP provides two core capabilities for AI coding agents:

    1. **Decision Audit Trail** -- every agent decision is logged with reasoning, confidence, and policy outcome.
    2. **Documentation Engine** -- auto-generates session summaries, risk reports, and pattern analysis from decision records.

    ## Response Formats

    The REST API (`adp-server`, port 8080) uses envelope wrappers for all responses:

    - **Single object**: `{"data": <object>}`
    - **Created object** (201): `{"data": <object>}`
    - **List**: `{"items": [...], "total": 10, "limit": 50, "offset": 0}`
    - **Error**: `{"error": "message", "code": "bad_request", "details": null}`

    ## Sidecar vs REST API

    The commit endpoints (`/v1/commits/*`) exist in two places with **different response formats**:

    - **REST API** (this spec, port 8080): Wraps responses in `{"data": {...}}` envelopes.
    - **MCP HTTP sidecar** (port 8081): Returns top-level JSON without envelopes. Git hooks expect the sidecar format.

    See the commit endpoint descriptions for details on both formats.

    ## Authentication

    ADP supports three authentication methods:

...

    The first registered user is automatically granted the admin role.
...

    ## Database Backends

...

...
servers:
...
paths:
...
  /v1/audit/decisions:
    post:
      operationId: logDecision
      summary: Log an agent decision
      description: |
        Records a decision made by an AI agent, including the action taken,
        reasoning, confidence score, and any alternatives considered.
        Optionally includes the policy evaluation result from a prior `checkAction` call.
      tags: [Audit]
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogDecisionRequest"
            example:
              session_id: "abc-123"
              decision_type: "file_modification"
              action: "Update nginx.conf proxy_pass directive"
              target:
                path: "/etc/nginx/nginx.conf"
                line: 42
              reasoning:
                rationale: "Current timeout causes 504 errors under load"
                evidence: "Error logs show 12 timeouts in last hour"
              confidence: 0.92
              alternatives:
                - action: "Increase worker connections instead"
                  reason: "Could help but doesn't address root cause"
                  confidence: 0.45
              policy_result:
                allowed: true
                denied_reasons: []
                policy_names: ["default"]
      responses:
        "201":
          description: Decision recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/DecisionRecord"
        "400":
          description: Missing session_id, decision_type, or action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      operationId: listDecisions
      summary: List decision records
      description: Returns a paginated list of decision records with optional filters.
      tags: [Audit]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: session_id
          in: query
          schema:
            type: string
          description: Filter by session ID
        - name: decision_type
          in: query
          schema:
            type: string
          description: Filter by decision type
        - name: status
          in: query
          schema:
            type: string
          description: Filter by decision status
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paginated list of decision records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/audit/decisions/{id}:
    get:
      operationId: getDecision
      summary: Get a decision record by ID
      tags: [Audit]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/DecisionId"
      responses:
        "200":
          description: Decision record details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/DecisionRecord"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/commits/prepare:
    post:
      operationId: prepareCommit
      summary: Prepare a commit for policy evaluation
      description: |
        Evaluates staged files against active policies and returns a commit token
        if approved. The commit token must be passed to `registerCommit` after
        the git commit succeeds.

        **Two response formats exist for this endpoint:**

        - **REST API** (`adp-server`, port 8080): Returns `{"data": {"commit_token": "...", "approved": true, "reason": "..."}}`.
        - **MCP HTTP sidecar** (`adp-mcp`, port 8081): Returns `{"commit_token": "...", "approved": true, "reason": "..."}` (no envelope). Git hooks expect this format.

        This specification documents the REST API envelope format.
      tags: [Commits]
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrepareCommitRequest"
            example:
              session_id: "abc-123"
              files:
                - "internal/api/handlers/sessions.go"
                - "internal/api/handlers/sessions_test.go"
              message: "feat(sessions): add TTL validation"
      responses:
        "201":
          description: Commit prepared and evaluated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/PrepareCommitResult"
              example:
                data:
                  commit_token: "ct_a1b2c3d4e5f6"
                  approved: true
                  reason: "Automatic approval"
        "400":
          description: Missing session_id or files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "session_id is required"
                code: "bad_request"
                details: null
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/commits/register:
    post:
      operationId: registerCommit
      summary: Register a completed commit
      description: |
        Associates a git commit SHA with a previously prepared commit token.
        Called by the `post-commit` hook after `git commit` succeeds.

        **Two response formats exist for this endpoint:**

        - **REST API** (port 8080): Returns `{"data": <CommitRecord>}`.
        - **MCP HTTP sidecar** (port 8081): Returns `{"status": "...", "commit_sha": "..."}` (no envelope).
      tags: [Commits]
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterCommitRequest"
            example:
              commit_token: "ct_a1b2c3d4e5f6"
              commit_sha: "a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2"
      responses:
        "200":
          description: Commit registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/CommitRecord"
        "400":
          description: Missing commit_token or commit_sha
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Commit token not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Commit token not found"
                code: "not_found"
                details: null
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/commits/verify:
    post:
      operationId: verifyCommit
      summary: Verify a commit has a complete audit trail
      description: |
        Checks whether a commit SHA has a valid audit trail (was prepared and
        registered through ADP). Called by the `pre-push` hook to ensure all
        commits being pushed were properly tracked.

        **Two response formats exist for this endpoint:**

        - **REST API** (port 8080): Returns `{"data": {"verified": true, "commit_sha": "...", "session_id": "...", "verified_at": "..."}}`.
        - **MCP HTTP sidecar** (port 8081): Returns `{"verified": true, "commit_sha": "..."}` (no envelope).

        **Backend note:** SQLite `IsCommitVerified` accepts both "committed" and
        "verified" status. PostgreSQL has a separate `Verify()` method that sets
        a `verified_at` timestamp.
      tags: [Commits]
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyCommitRequest"
            example:
              commit_sha: "a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2"
      responses:
        "200":
          description: Commit verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/VerifyCommitResult"
              example:
                data:
                  verified: true
                  commit_sha: "a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2"
                  session_id: "abc-123"
                  verified_at: "2026-02-20T14:30:00Z"
        "400":
          description: Missing commit_sha
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # Policies
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT authentication via locally-generated tokens (POST /v1/auth/login)
        or external JWKS provider. Required in production (PostgreSQL mode).
        Optional in SQLite development mode.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Shared API key authentication. Set via ADP_API_KEY environment variable.

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    DecisionId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Decision record ID
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 500
      description: Maximum number of items to return
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Number of items to skip for pagination

  # ---------------------------------------------------------------------------
  # Reusable Responses
  # ---------------------------------------------------------------------------
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "not found"
            code: "not_found"
            details: null
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "authentication required"
            code: "unauthorized"
            details: null
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "internal server error"
            code: "internal_error"
            details: null

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
          enum:
            - bad_request
            - not_found
            - unauthorized
            - forbidden
            - internal_error
            - conflict
        details:
          description: Additional error context (nullable)
          oneOf:
            - type: object
            - type: "null"

    # -------------------------------------------------------------------------
    # Sessions
    # -------------------------------------------------------------------------
...
    LogDecisionRequest:
      type: object
      required: [session_id, decision_type, action]
      properties:
        session_id:
          type: string
          description: Active session ID
        decision_type:
          type: string
          description: Category of decision (e.g., file_modification, shell_command, architecture)
        action:
          type: string
          description: Human-readable description of the action taken
        target:
          type: object
          additionalProperties: true
          description: What the action targets (file path, resource, etc.)
        reasoning:
          type: object
          additionalProperties: true
          description: Structured reasoning behind the decision
        confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
          default: 0.8
          description: Agent's confidence in this decision (0.0 to 1.0)
        alternatives:
          type: array
          items:
            $ref: "#/components/schemas/Alternative"
          description: Other options the agent considered
        context_snapshot:
          type: object
          additionalProperties: true
          description: Snapshot of relevant context at decision time
        policy_result:
          $ref: "#/components/schemas/PolicyResult"
    Alternative:
      type: object
      properties:
        action:
          type: string
          description: Alternative action description
        reason:
          type: string
          description: Why this alternative was not chosen
        confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: Confidence in this alternative
    PolicyResult:
      type: object
      properties:
        allowed:
          type: boolean
          description: Whether policy evaluation allowed the action
        denied_reasons:
          type: array
          items:
            type: string
          description: Reasons for denial
        policy_names:
          type: array
          items:
            type: string
          description: Policies that were evaluated
    DecisionRecord:
      type: object
      properties:
        id:
          type: string
          description: Unique decision record ID
        session_id:
          type: string
        decision_type:
          type: string
        action:
          type: string
        target:
          type: object
          additionalProperties: true
        reasoning:
          type: object
          additionalProperties: true
        confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
        alternatives:
          type: array
          items:
            $ref: "#/components/schemas/Alternative"
        context_snapshot:
          type: object
          additionalProperties: true
        policy_result:
          $ref: "#/components/schemas/PolicyResult"
        status:
          type: string
          description: Decision execution status
        outcome:
          type: string
          description: Outcome after execution (success, failure, etc.)
        created_at:
          type: string
          format: date-time
    DecisionListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DecisionRecord"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # -------------------------------------------------------------------------
    # Commits
    # -------------------------------------------------------------------------
    PrepareCommitRequest:
      type: object
      required: [session_id, files]
      properties:
        session_id:
          type: string
          description: Active session ID
        files:
          type: array
          items:
            type: string
          minItems: 1
          description: List of staged file paths
        message:
          type: string
          description: Intended commit message
    PrepareCommitResult:
      type: object
      properties:
        commit_token:
          type: string
          description: Token to use when registering the commit after `git commit`
        approved:
          type: boolean
          description: Whether the commit is approved by policy evaluation
        reason:
          type: string
          description: Reason for approval or denial
    RegisterCommitRequest:
      type: object
      required: [commit_token, commit_sha]
      properties:
        commit_token:
          type: string
          description: Token from a prior `prepareCommit` call
        commit_sha:
          type: string
          description: Git commit SHA from the completed commit
    CommitRecord:
      type: object
      properties:
        id:
          type: string
          description: Unique commit record ID
        session_id:
          type: string
        commit_sha:
          type: string
          description: Git commit SHA
        commit_token:
          type: string
          description: Preparation token
        files:
          type: array
          items:
            type: string
          description: Files included in the commit
        message:
          type: string
          description: Commit message
        status:
          type: string
          description: Commit record status (prepared, committed, verified)
        approved:
          type: boolean
          description: Whether the commit was approved by policy
        approval_reason:
          type: string
          description: Reason for approval decision
        prepared_at:
          type: string
          format: date-time
        committed_at:
          type: string
          format: date-time
          description: When the commit was registered
        verified_at:
          type: string
          format: date-time
          description: When the commit was verified (PostgreSQL only; SQLite infers from status)
    VerifyCommitRequest:
      type: object
      required: [commit_sha]
      properties:
        commit_sha:
          type: string
          description: Git commit SHA to verify
    VerifyCommitResult:
      type: object
      properties:
        verified:
          type: boolean
          description: Whether the commit has a complete audit trail
        commit_sha:
          type: string
        session_id:
          type: string
          description: Session that produced this commit
        verified_at:
          type: string
          format: date-time
          description: Verification timestamp

    # -------------------------------------------------------------------------
    # Policies
    # -------------------------------------------------------------------------
...